refactor: interfaceをtypeに変更
汎用性向上のため
